refactor(write): share one buildFrontmatter between create and update
The two routes carried identical copies, so adding a field meant editing
both — canonical was added to only one and silently dropped on save.
